refactor(stella-tui): cut the tool-call block out of render/entry.rs
`crates/stella-tui/src/render/entry.rs` was 1246 lines with no entry in
`scripts/file-size-baseline.txt`, so the `file-size` guard judged it against
the flat 1500-line ceiling — 83% of a hard gate, on the file that grows
whenever a transcript entry kind gains a rendering decision (#4019, #4020,
#4036, #4123, #4155, #4169, #4210 all landed in it). The next feature to add
~280 lines would have discovered that mid-PR and paid for a split under time
pressure inside a feature diff, which is the shape AGENTS.md's god-file rule
exists to prevent.

The cut follows `render/entry/recall.rs`'s precedent — a concern, not a line
count. The tool-call block stacks four independent rendering decisions
(syntax highlighting in the body's own language, the emitter's line-number
gutter, a preview window anchored on a salient line, and a truncation notice
that must agree with what the export and Observatory surfaces fold), and it
is the densest arm of `entry_body`.

Moved to `render/entry/tool.rs`: `OK_PREVIEW`, `push_body_line`,
`argument_rows`, and the `TranscriptEntry::ToolResult` arm's body. Nothing
else referenced any of them.

No behaviour changes. Two signatures moved rather than copied, both named in
the new module's doc: `result_body` takes the whole `TranscriptEntry` instead
of seven destructured fields, because seven plus the view and three render
parameters is over `clippy::too_many_arguments` and the honest answer to that
lint is fewer arguments rather than an `#[allow]`; and `argument_rows` spells
its first parameter `Color` under this module's imports.

entry.rs 1246 -> 906. No baseline entry added, and no ceiling raised.

Refs #4230
